Solucionado definitivamente el error de añadir más de 1 producto con el mismo id al carrito y solucionado el error de poder añadir varios carritos en la pestaña de productos, así como añadido un nuevo producto

+++ IberComponents/src/main/java/es/code/urjc/ibercomponents/entities/Product.java
@@ -37,7 +37,7 @@ public class Product
     private List<Review> reviews;
 
 
-    public Product(String name, String descipcion, double d, int i, String features, boolean imageBool) {
+    public Product(String name, String descipcion, double d, double i, String features, boolean imageBool) {
         this.name = name;
         this.description = descipcion;
         this.score = d;
